feat: Implementa alocação de escala via ORM e integração com trigger de sobreposição

+++ hospital-management/src/main/java/ufpb/hospital/model/Escala.java
@@ -20,7 +20,7 @@ public class Escala {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_escala")
-    private Long idEscala;
+    private Integer idEscala;
 
     @ManyToOne
     @JoinColumn(name = "id_unidade", nullable = false)
